#96 rename format meta model package to project
