+++ api/common/src/main/scala/de/htwg/zeta/common/format/model/GraphicalDslInstanceFormat.scala
@@ -2,9 +2,9 @@ package de.htwg.zeta.common.format.model
 
 import java.util.UUID
 
-import de.htwg.zeta.common.format.metaModel.AttributeFormat
-import de.htwg.zeta.common.format.metaModel.AttributeValueFormat
-import de.htwg.zeta.common.format.metaModel.MethodFormat
+import de.htwg.zeta.common.format.project.AttributeFormat
+import de.htwg.zeta.common.format.project.AttributeValueFormat
+import de.htwg.zeta.common.format.project.MethodFormat
 import de.htwg.zeta.common.models.project.instance
 import de.htwg.zeta.common.models.project.instance
 import de.htwg.zeta.common.models.project.instance.GraphicalDslInstance
