+++ api/common/src/main/scala/de/htwg/zeta/common/format/project/GraphicalDslReleaseFormat.scala
@@ -1,4 +1,4 @@
-package de.htwg.zeta.common.format.metaModel
+package de.htwg.zeta.common.format.project
 
 import java.util.UUID
