+++ api/server/app/de/htwg/zeta/server/module/JsonFormatModule.scala
@@ -10,16 +10,16 @@ import de.htwg.zeta.common.format.entity.FilterFormat
 import de.htwg.zeta.common.format.entity.GeneratorFormat
 import de.htwg.zeta.common.format.entity.GeneratorImageFormat
 import de.htwg.zeta.common.format.entity.TimedTaskFormat
-import de.htwg.zeta.common.format.metaModel.AttributeFormat
-import de.htwg.zeta.common.format.metaModel.AttributeTypeFormat
-import de.htwg.zeta.common.format.metaModel.AttributeValueFormat
-import de.htwg.zeta.common.format.metaModel.ClassFormat
-import de.htwg.zeta.common.format.metaModel.ConceptFormat
-import de.htwg.zeta.common.format.metaModel.EnumFormat
-import de.htwg.zeta.common.format.metaModel.GdslProjectFormat
-import de.htwg.zeta.common.format.metaModel.GraphicalDslReleaseFormat
-import de.htwg.zeta.common.format.metaModel.MethodFormat
-import de.htwg.zeta.common.format.metaModel.ReferenceFormat
+import de.htwg.zeta.common.format.project.AttributeFormat
+import de.htwg.zeta.common.format.project.AttributeTypeFormat
+import de.htwg.zeta.common.format.project.AttributeValueFormat
+import de.htwg.zeta.common.format.project.ClassFormat
+import de.htwg.zeta.common.format.project.ConceptFormat
+import de.htwg.zeta.common.format.project.EnumFormat
+import de.htwg.zeta.common.format.project.GdslProjectFormat
+import de.htwg.zeta.common.format.project.GraphicalDslReleaseFormat
+import de.htwg.zeta.common.format.project.MethodFormat
+import de.htwg.zeta.common.format.project.ReferenceFormat
 import de.htwg.zeta.common.format.model.EdgeFormat
 import de.htwg.zeta.common.format.model.GraphicalDslInstanceFormat
 import de.htwg.zeta.common.format.model.NodeFormat
